fix(): Return error code if dns resolution fails
The kubeslice-dns plugin does not return an error code if the resolution
fails due to unsupported query type or if records are not found for the
domain name. It returns a success code instead with an empty Answers
section. This can cause certain dns client and dns proxies to
misinterpret the response and to cache responses with empty Answers.

Added this fix to return appropriate error codes.

Signed-off-by: Bharath Horatti <bharath@aveshasystems.com>